src/requirements.txt
pip==22.3
google==3.0.0
google-api-core==2.8.2
google-api-python-client==2.58.0
google-cloud-bigquery==3.3.2
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-auth==2.11.0
# google-cloud==0.34.0
PyYAML==5.4.1
# protobuf==3.15.6

# scipy==1.6.1
# xgboost==1.3.3
scikit-learn==1.0.2
pandas==1.3.5
numpy==1.21.6

six==1.16
google-resumable-media==2.3.3


# to be able to call pandas.read_csv for a file in Storage
fsspec==2022.7.1
gcsfs==2022.7.1

protobuf~=3.19.0
